ENH: rework visualization module
