+++ docs/modules/visualization.rst
@@ -4,47 +4,19 @@ Visualization
 The visualization module provides a variety of functions to visualize initial, 
 intermediate and final datasets within trackintel. 
 
-Positionfixes
+plot_map
 =============
 
-.. autofunction:: trackintel.visualization.positionfixes.plot_positionfixes
-
-.. image:: /_static/example_positionfixes.png
-   :scale: 25 %
-   :align: center
-
-Staypoints
-==========
-
-.. autofunction:: trackintel.visualization.staypoints.plot_staypoints
-
-.. image:: /_static/example_staypoints.png
-   :scale: 25 %
-   :align: center
-
-Triplegs
-========
-
-.. autofunction:: trackintel.visualization.triplegs.plot_triplegs
-
-.. image:: /_static/example_triplegs.png
-   :scale: 25 %
-   :align: center
-
-Locations
-=========
-
-.. autofunction:: trackintel.visualization.locations.plot_locations
+.. autofunction:: trackintel.visualization.plot_map
 
 .. image:: /_static/example_locations.png
    :scale: 25 %
    :align: center
 
-
 Modal split
 ===========
 
-.. autofunction:: trackintel.visualization.modal_split.plot_modal_split
+.. autofunction:: trackintel.visualization.plot_modal_split
 
 .. image:: /_static/example_modal_split_geolife.png
    :scale: 18 %
