+++ trackintel/model/staypoints.py
@@ -7,7 +7,6 @@ from trackintel.io.postgis import write_staypoints_postgis
 from trackintel.model.util import _copy_docstring, _register_trackintel_accessor
 from trackintel.preprocessing.filter import spatial_filter
 from trackintel.preprocessing.staypoints import generate_locations, merge_staypoints
-from trackintel.visualization.staypoints import plot_staypoints
 
 
 @_register_trackintel_accessor("as_staypoints")
@@ -112,15 +111,6 @@ class StaypointsAccessor(object):
         See :func:`trackintel.preprocessing.filter.spatial_filter`.
         """
         return ti.preprocessing.filter.spatial_filter(self._obj, *args, **kwargs)
-
-    @_copy_docstring(plot_staypoints)
-    def plot(self, *args, **kwargs):
-        """
-        Plot this collection of staypoints.
-
-        See :func:`trackintel.visualization.staypoints.plot_staypoints`.
-        """
-        ti.visualization.staypoints.plot_staypoints(self._obj, *args, **kwargs)
 
     @_copy_docstring(write_staypoints_csv)
     def to_csv(self, filename, *args, **kwargs):
