+++ trackintel/visualization/__init__.py
@@ -1,16 +1,10 @@
-from .locations import plot_locations
-from .osm import plot_osm_streets
-from .positionfixes import plot_positionfixes
-from .staypoints import plot_staypoints
-from .triplegs import plot_triplegs
-from .modal_split import plot_modal_split
+from .plotting import plot_map
+from .plotting import plot_osm_streets
+from .plotting import plot_modal_split
 
 
 __all__ = [
-    "plot_locations",
+    "plot_map",
     "plot_osm_streets",
-    "plot_positionfixes",
-    "plot_staypoints",
-    "plot_triplegs",
     "plot_modal_split",
 ]
